\name{bin_UC_models}
\alias{bin_UC_models}
\title{This function take a UCsims object and runs a set of logistic models on the data
if coeffs == TRUE, it returns a list of model coefficients
if coef == FALSE, it returns a list of glm model objects}
\usage{
bin_UC_models(UCsim_obj, coefs = TRUE, ...)
}
\description{
This function take a UCsims object and runs a set of
logistic models on the data if coeffs == TRUE, it returns a
list of model coefficients if coef == FALSE, it returns a
list of glm model objects
}

