% Generated by roxygen2 (4.0.1): do not edit by hand
\name{catch_collinear}
\alias{catch_collinear}
\title{wrapper function for simulate_UC to stop any perfectly collinear datasets being generated}
\usage{
catch_collinear(orUE, BC, MC, subgroups, ...)
}
\description{
wrapper function for simulate_UC to stop any perfectly collinear datasets being generated
}

