% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lin_adjust}
\alias{lin_adjust}
\title{Performs the adjustment for unmeasured confounding according to Lin}
\usage{
lin_adjust(sims_obj, BC = "U1", MC = NULL, subgroups = NULL)
}
\description{
Performs the adjustment for unmeasured confounding according to Lin
}

