\name{overall_prev}
\alias{overall_prev}
\title{This function uses an iterative algorithm to obtain overall prevalences of Y and E.}
\usage{
overall_prev(sims_fn = lin_sim, iter_fn = iter_fn_1, pY_target, pE_target,
  probE_0 = 0.2, probY_0 = 0.7, tol_Y = 5e-04, tol_E = 5e-04,
  verbose = TRUE, ...)
}
\description{
This function uses an iterative algorithm to obtain overall
prevalences of Y and E.
}

