% Generated by roxygen2 (4.0.1): do not edit by hand
\name{r_and_r_adjust}
\alias{r_and_r_adjust}
\title{performs the R and R sensitivity on the subgroups of the stratified measured confounder}
\usage{
r_and_r_adjust(sims_obj, subgroups = 5, BC = "U1", MC = "U2")
}
\description{
Note that R&R are wrong in saying that alpha and gamma are for E=0; they are actually for E=1
To get the log odds and log odds-ratio for E=0 alpha and gamma need to be inverted
(because they are in logs, this is done by changing the signs)
}

