% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{rUCsims}
\alias{rUCsims}
\alias{rUCsims-package}
\title{rUCsims: Functions for simulating and accounting for unmeasured confounding}
\description{
rUCsims: Functions for simulating and accounting for unmeasured confounding
}

