% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_sims}
\alias{run_sims}
\title{Helper function to automate running different conditions}
\usage{
run_sims(Ubs, aggregator = median, rawdata = TRUE, agg_subset = NULL, ...)
}
\description{
Helper function to automate running different conditions
}

