% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simulate_UC}
\alias{simulate_UC}
\title{This function generates the simulated exposure and response data, based on a binary and an
optional continuous confounder.}
\usage{
simulate_UC(obs, orUE, probE_0, probY_0, orEY_0, probU, modUB, orME, orMY_0,
  rho, orUEy_0 = NULL, method = c("continuous_confounder",
  "no_continuous_confounder"))
}
\arguments{
\item{obs}{the number of observations in the simulated data (number of rows)}

\item{orUE}{the odds ratio OR(ue)}

\item{probE_0}{probability that E=1 when (all confounders)= 0}

\item{probY_0}{p(y=1/exposure=0 & confounders=0)}

\item{orEY_0}{odds ratio OR(ey)}

\item{probU}{p(u=1) prevalence of the binary unmeasured confounder}

\item{modUB}{moderating effect of the UC on OR(ey)}

\item{orME}{odds ratio OR(me)}

\item{orMY_0}{odds ratio OR(my/e=0)}

\item{rho}{correlation coefficient between U1 and U2}

\item{orUEy_0}{}

\item{method}{}
}
\value{
object of class 'UCsims' which is a list containing a dataframe of the simulated data, a vector of parameter values and the method.
}
\description{
The exposure is generated as a function of the confounders and the response is then generated
as a function of the exposure and the confounders
}
\details{
The binary confounder is labeled "U1" and the continuous "U2".
Confounders can be correlated, using the `rho` argument.
Linear moderation can also be introduced via the `modUB` argument between the continuous
confounder and the exposure/outcome.

The method argument specifies how the exposure (E) and outcome (Y) variables are generated.
The "continuous_confounder" option means that E and Y are based on both the binary and
continuous confounder. The "no_continuous_confounder" option means that E and Y are a function
of just the binary confounder.  If this is selected, the unmeasured confounder U2 is still generated
and may be correlated with U1, although it will be uncorrelated with both E and Y.
}

