\name{catch_collinear}
\alias{catch_collinear}
\title{wrapper function for the sims_fn to stop any perfectly collinear datasets being generated}
\usage{
catch_collinear(orUB, sims_fn, UC, ...)
}
\description{
wrapper function for the sims_fn to stop any perfectly
collinear datasets being generated
}

