% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_UC_levels}
\alias{run_UC_levels}
\title{wrapper function for adjust functions
Runs both functions over a range of orUE levels
returns a dataframe of parameter values from the simulation and the models
calculates the adjustments for binary and continuous UCs}
\usage{
run_UC_levels(orUE_levels, method = "no_moderator", adjust_fn = lin_adjust,
  BC = "U1", MC = "U2", subgroups = 5, ...)
}
\description{
wrapper function for adjust functions
Runs both functions over a range of orUE levels
returns a dataframe of parameter values from the simulation and the models
calculates the adjustments for binary and continuous UCs
}

